+++ include/render/primitives/MeshPrimitive.h
@@ -71,8 +71,8 @@ namespace render {
                                         const TransformedLeafVisitor& visitor) const override;
     void appendIntersectionSceneRecords(IntersectionSceneBuilder& builder,
                                         std::shared_ptr<render::Material> inheritedMaterial,
-                                        const Matrix4d& pointMatrix,
-                                        const Matrix3d& normalMatrix) const override;
+                                        const Matrix4d& pointMatrix, const Matrix3d& normalMatrix,
+                                        const Primitive* inheritedObject = nullptr) const override;
 
     std::shared_ptr<Mesh> tessellate(int lod = 0) const override;